+++ app/uk/gov/hmrc/disareturnsstubs/repositories/generatereport/ReportIssueRepository.scala
@@ -45,7 +45,7 @@ class ReportIssueRepository @Inject() (mc: MongoComponent, appConfig: AppConfig)
           Indexes.ascending("createdAt"),
           IndexOptions()
             .name("createdAt_ttl_index")
-            .expireAfter(appConfig.reportTtlDays, TimeUnit.DAYS)
+            .expireAfter(appConfig.reportTtlDays.toLong, TimeUnit.DAYS)
         )
       )
     )
